list and ralist StressTesters cleaned up

+++ src/test/java/org/javimmutable/collections/StressTestTool/JImmutableListStressTester.java
@@ -8,6 +8,12 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Random;
 
+/**
+ * Test program for all implementations of JImmutableList, including those that implement
+ * JImmutableRandomAccessList. Divided into three sections: growing (adds new values to
+ * either end of the list), shrinking (removes values from either end), and cleanup (empties
+ * the list of all values)
+ */
 public class JImmutableListStressTester
         extends AbstractListStressTestable
 {
@@ -29,6 +35,7 @@ public class JImmutableListStressTester
     public void execute(Random random,
                         JImmutableList<String> tokens)
     {
+        JImmutableList<String> list = this.list;
         List<String> expected = new ArrayList<String>();
         int size = random.nextInt(100000);
         System.out.printf("JImmutableListStressTest on %s of size %d%n", list.getClass().getSimpleName(), size);
@@ -36,84 +43,86 @@ public class JImmutableListStressTester
         for (int loops = 1; loops <= 6; ++loops) {
             System.out.printf("growing %d%n", list.size());
             for (int i = 0; i < size / 3; ++i) {
-                if (random.nextBoolean()) {
+                switch (random.nextInt(11)) {
+                case 0: //assign(int, T)
+                    if (!list.isEmpty()) {
+                        String value = makeValue(tokens, random);
+                        int index = random.nextInt(list.size());
+                        list = list.assign(index, value);
+                        expected.set(index, value);
+                        break;
+                    }
+                case 1: //insert(T)
                     String value = makeValue(tokens, random);
-                    switch (random.nextInt(3) + 1) {
-                    case 0:
-                        if (list.isEmpty()) {
-                            break;
-                        } else {
-                            int index = random.nextInt(list.size());
-                            list = list.assign(index, value);
-                            expected.set(index, value);
-                            break;
-                        }
-                    case 1:
-                        list = list.insert(value);
-                        expected.add(value);
-                        break;
-                    case 2:
-                        list = list.insertFirst(value);
-                        expected.add(0, value);
-                        break;
-                    case 3:
-                        list = list.insertLast(value);
-                        expected.add(value);
-                        break;
-                    default:
-                        throw new RuntimeException();
-                    }
-                } else {
-                    List<String> values = new ArrayList<String>();
-                    for (int n = 0; n < random.nextInt(3); ++n) {
-                        values.add(makeValue(tokens, random));
-                    }
-                    switch (random.nextInt(7)) {
-                    case 0:
-                        list = list.insert(values);
-                        expected.addAll(values);
-                        break;
-                    case 1:
-                        list = list.insertAll(IterableCursorable.of(values));
-                        expected.addAll(values);
-                        break;
-                    case 2:
-                        list = list.insertAll(values);
-                        expected.addAll(values);
-                        break;
-                    case 3:
-                        list = list.insertAllLast(IterableCursorable.of(values));
-                        expected.addAll(values);
-                        break;
-                    case 4:
-                        list = list.insertAllLast(values);
-                        expected.addAll(values);
-                        break;
-                    case 5:
-                        list = list.insertAllFirst(IterableCursorable.of(values));
-                        expected.addAll(0, values);
-                        break;
-                    case 6:
-                        list = list.insertAllFirst(values);
-                        expected.addAll(0, values);
-                        break;
-                    default:
-                        throw new RuntimeException();
-                    }
+                    list = list.insert(value);
+                    expected.add(value);
+                    break;
+                case 2: //insertFirst(T)
+                    value = makeValue(tokens, random);
+                    list = list.insertFirst(value);
+                    expected.add(0, value);
+                    break;
+                case 3: //insertLast(T)
+                    value = makeValue(tokens, random);
+                    list = list.insertLast(value);
+                    expected.add(value);
+                    break;
+                case 4: //insert(Iterable)
+                    List<String> values = makeInsertList(tokens, random);
+                    list = list.insert(values);
+                    expected.addAll(values);
+                    break;
+                case 5: //insert(Cursorable)
+                    values = makeInsertList(tokens, random);
+                    list = list.insertAll(IterableCursorable.of(values));
+                    expected.addAll(values);
+                    break;
+                case 6: //insertAll(Collection)
+                    values = makeInsertList(tokens, random);
+                    list = list.insertAll(values);
+                    expected.addAll(values);
+                    break;
+                case 7: //insertAllLast(Cursorable)
+                    values = makeInsertList(tokens, random);
+                    list = list.insertAllLast(IterableCursorable.of(values));
+                    expected.addAll(values);
+                    break;
+                case 8: //insertAllLast(Collection)
+                    values = makeInsertList(tokens, random);
+                    list = list.insertAllLast(values);
+                    expected.addAll(values);
+                    break;
+                case 9: //insertAllFirst(Cursorable)
+                    values = makeInsertList(tokens, random);
+                    list = list.insertAllFirst(IterableCursorable.of(values));
+                    expected.addAll(0, values);
+                    break;
+                case 10: //insertAllFirst(Collection)
+                    values = makeInsertList(tokens, random);
+                    list = list.insertAllFirst(values);
+                    expected.addAll(0, values);
+                    break;
+                default:
+                    throw new RuntimeException();
+
                 }
             }
-            verifyContents(expected, list);
+            verifyContents(list, expected);
             System.out.printf("shrinking %d%n", list.size());
             for (int i = 0; i < size / 6; ++i) {
-                if (random.nextBoolean()) {
+                switch (random.nextInt(2)) {
+                case 0: //deleteLast()
                     list = list.deleteLast();
                     expected.remove(expected.size() - 1);
-                } else {
+                    break;
+                case 1: //deleteFirst()
                     list = list.deleteFirst();
                     expected.remove(0);
+                    break;
                 }
             }
-            verifyContents(expected, list);
+            verifyContents(list, expected);
+            verifyCursor(list, expected);
         }
         System.out.printf("cleanup %d%n", expected.size());
         while (list.size() > random.nextInt(3)) {
@@ -125,13 +134,16 @@ public class JImmutableListStressTester
                 expected.remove(0);
             }
         }
-        verifyContents(expected, list);
-        list = list.deleteAll();
-        expected.clear();
+        if (list.size() != 0) {
+            verifyContents(list, expected);
+            list = list.deleteAll();
+            expected.clear();
+        }
         if (list.size() != 0) {
             throw new RuntimeException(String.format("expected map to be empty but it contained %d keys%n", list.size()));
         }
-        verifyContents(expected, list);
+        verifyContents(list, expected);
         System.out.printf("JImmutableListStressTest on %s completed without errors%n", list.getClass().getSimpleName());
     }
 }
+
